cleanup and improve errors
