+++ src/errors/line_error.rs
@@ -2,7 +2,6 @@ use std::fmt;
 
 use pyo3::prelude::*;
 use pyo3::types::PyDict;
-use strum::EnumMessage;
 
 use super::kinds::ErrorKind;
 
@@ -55,139 +54,6 @@ impl ValLineError {
             new.location = [location.clone(), new.location].concat();
         }
         new
-    }
-
-    pub fn into_py_error(self) -> PyLineError {
-        PyLineError::new(self)
-    }
-}
-
-// impl IntoPy<PyLineError> for ValLineError {
-//     fn into_py(self, py: Python) -> PyLineError {
-//         PyLineError::new(self).to_object(py)
-//     }
-// }
-//
-// impl FromPyObject for PyLineError {
-//     fn extract(ob: PyAny) -> PyResult<Self> {
-//         todo!()
-//         // LineError::new(self).to_object(py)
-//     }
-// }
-
-/// `PyLineError` are the public version of `ValLineError`, as help and used in `ValidationError`s
-#[pyclass]
-#[derive(Debug, Clone)]
-pub struct PyLineError {
-    kind: ErrorKind,
-    location: Location,
-    message: Option<String>,
-    input_value: Option<PyObject>,
-    context: Option<Context>,
-}
-
-impl PyLineError {
-    pub fn new(raw_error: ValLineError) -> Self {
-        Self {
-            kind: raw_error.kind,
-            location: raw_error.location,
-            message: raw_error.message,
-            input_value: raw_error.input_value,
-            context: raw_error.context,
-        }
-    }
-
-    pub fn as_dict(&self, py: Python) -> PyResult<PyObject> {
-        let dict = PyDict::new(py);
-        dict.set_item("kind", self.kind())?;
-        dict.set_item("loc", self.location(py))?;
-        dict.set_item("message", self.message())?;
-        if let Some(input_value) = &self.input_value {
-            dict.set_item("input_value", input_value)?;
-        }
-        if let Some(context) = &self.context {
-            dict.set_item("context", context)?;
-        }
-        Ok(dict.into_py(py))
-    }
-
-    fn kind(&self) -> String {
-        self.kind.to_string()
-    }
-
-    fn location(&self, py: Python) -> PyObject {
-        let mut loc: Vec<PyObject> = Vec::with_capacity(self.location.len());
-        for location in &self.location {
-            let item: PyObject = match location {
-                LocItem::S(key) => key.into_py(py),
-                LocItem::I(index) => index.into_py(py),
-            };
-            loc.push(item);
-        }
-        loc.into_py(py)
-    }
-
-    fn message(&self) -> String {
-        let raw = self.raw_message();
-        match self.context {
-            Some(ref context) => context.render(raw),
-            None => raw,
-        }
-    }
-
-    fn raw_message(&self) -> String {
-        // TODO string substitution
-        if let Some(ref message) = self.message {
-            message.to_string()
-        } else {
-            match self.kind.get_message() {
-                Some(message) => message.to_string(),
-                None => self.kind(),
-            }
-        }
-    }
-
-    fn format_pretty(&self, f: &mut fmt::Formatter<'_>, py: Option<Python>) -> fmt::Result {
-        if !self.location.is_empty() {
-            let loc = self
-                .location
-                .iter()
-                .map(|i| i.to_string())
-                .collect::<Vec<String>>()
-                .join(" -> ");
-            writeln!(f, "{}", loc)?;
-        }
-        write!(f, "  {} [kind={}", self.message(), self.kind())?;
-        if let Some(ctx) = &self.context {
-            write!(f, ", context={}", ctx)?;
-        }
-        if let Some(input_value) = &self.input_value {
-            write!(f, ", input_value={}", input_value)?;
-            if let Some(py) = py {
-                if let Ok(type_) = input_value.as_ref(py).get_type().name() {
-                    write!(f, ", input_type={}", type_)?;
-                }
-            }
-        }
-        write!(f, "]")
-    }
-
-    pub fn pretty(&self, py: Option<Python>) -> String {
-        format!("{}", Fmt(|f| self.format_pretty(f, py)))
-    }
-}
-
-// hack from https://users.rust-lang.org/t/reusing-an-fmt-formatter/8531/4
-struct Fmt<F>(pub F)
-where
-    F: Fn(&mut fmt::Formatter) -> fmt::Result;
-
-impl<F> fmt::Display for Fmt<F>
-where
-    F: Fn(&mut fmt::Formatter) -> fmt::Result,
-{
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        (self.0)(f)
     }
 }
